ether_ioctl: Command type is u_long not int

+++ sys/net/if_ethersubr.c
@@ -602,7 +602,7 @@ ether_ifdetach(struct ifnet *ifp)
 }
 
 int
-ether_ioctl(struct ifnet *ifp, int command, caddr_t data)
+ether_ioctl(struct ifnet *ifp, u_long command, caddr_t data)
 {
 	struct ifaddr *ifa = (struct ifaddr *) data;
 	struct ifreq *ifr = (struct ifreq *) data;
