+++ sys/net/if_var.h
@@ -748,7 +748,7 @@ void	ether_input_chain(struct ifnet *, struct mbuf *,
 void	ether_input_chain_init(struct mbuf_chain *);
 void	ether_input_dispatch(struct mbuf_chain *);
 int	ether_output_frame(struct ifnet *, struct mbuf *);
-int	ether_ioctl(struct ifnet *, int, caddr_t);
+int	ether_ioctl(struct ifnet *, u_long, caddr_t);
 struct ifnet *ether_bridge_interface(struct ifnet *ifp);
 uint32_t	ether_crc32_le(const uint8_t *, size_t);
 uint32_t	ether_crc32_be(const uint8_t *, size_t);
